feat: Переиспользовала Navigation в Footer, добавила Scroll to top в Content
