+++ src/components/Footer/Footer.css
@@ -61,44 +61,34 @@
   max-width: 546px;
 }
 
-.footer__menu {
+/* .navigation_type_footer {
   margin: 0;
   padding: 0;
   grid-area: menu;
-}
-
-.footer__menu-item {
-  list-style: none;
-  padding: 0;
-  margin: 0 0 23px 0;
-}
-
-.footer__menu-link_none-decoration {
-  text-decoration: none;
-}
-
-.footer__menu-link {
-  padding: 0;
-  margin: 0;
-  line-height: 1.25;
-  font-size: 16px;
-  font-weight: normal;
-  opacity: 1;
-  transition: color 0.1s ease-in;
-  color: #faf1e9;
-}
-
-.footer__menu-link:hover {
-  color: #513629;
-}
-
-.footer__menu-links {
-  margin: 0;
-  padding: 0;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
 }
+
+a.navigation__link_underlined {
+  text-decoration: underline;
+}
+
+.navigation__link_type_footer {
+  padding: 0;
+  margin: 0 0 23px 0;
+  line-height: 1.25;
+  font-size: 16px;
+  font-weight: normal;
+  opacity: 1;
+  transition: color 0.26s ease-in;
+  color: #faf1e9;
+  text-decoration: none;
+}
+
+.navigation__link_type_footer:hover {
+  color: #513629;
+} */
 
 .footer__social {
   margin: 0 0 22px 0;
@@ -228,17 +218,20 @@
     justify-items: center;
   }
 
-  .footer__copyright {
+  .footer__wrapper {
     justify-content: center;
     align-items: center;
+  }
+
+  .footer__copyright {
     text-align: center;
     font-size: 8px;
     margin: 0;
   }
 
-  .footer__menu-links {
+  /* .navigation_type_footer {
     align-items: center;
-  }
+  } */
 
   .footer__top {
     justify-content: center;
@@ -251,17 +244,12 @@
   }
 }
 
-@media screen and (max-width: 480px) {
-  .footer__wrapper {
-    justify-content: center;
-    align-items: center;
-  }
-
-  .footer__menu-link {
+/* @media screen and (max-width: 480px) {
+  .navigation__link_type_footer {
     line-height: 1.2;
     font-size: 15px;
   }
-}
+} */
 
 @media screen and (max-width: 320px) {
   .footer__container {
