+++ src/components/Footer/texts.js
@@ -1,0 +1,6 @@
+// texts for footer
+// export const OFFER_TEXT = "Оферта";
+// export const TERMS_OF_USE_TEXT = "Политика конфиденциальности";
+export const COPYRIGHT_TEXT =
+  "Все фото и видео материалы являются собственностью автора. Любое копирование и использование материалов в сети Интернет или печатных СМИ, допускается только с условием размещения активной ссылки без запрета индексации или с письменного согласия администрации сайта.";
+export const TIME_TEXT = "пн - пт 09:00 - 21:00 по МСК";
