+++ src/config/texts.js
@@ -5,92 +5,5 @@ export const CATALOGUE_TEXT = 'Каталог';
 export const DELIVERY_TEXT = 'Доставка';
 export const CONTACTS_TEXT = 'Контакты';
 export const PHONE = '+7 925 343 34 34';
-
-// texts for footer
 export const OFFER_TEXT = "Оферта";
 export const TERMS_OF_USE_TEXT = "Политика конфиденциальности";
-export const COPYRIGHT_TEXT =
-  "Все фото и видео материалы являются собственностью автора. Любое копирование и использование материалов в сети Интернет или печатных СМИ, допускается только с условием размещения активной ссылки без запрета индексации или с письменного согласия администрации сайта.";
-export const TIME_TEXT = "пн - пт 09:00 - 21:00 по МСК";
-
-// texts for documents
-export const DOCUMENT_OFFER_TITLE = "Оферта";
-export const DOCUMENT_OFFER_CONTENT = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Nisl tincidunt eget nullam non. Quis hendrerit dolor magna eget est lorem ipsum dolor sit. Volutpat odio facilisis mauris sit amet massa. Commodo odio aenean sed adipiscing diam donec adipiscing tristique. Mi eget mauris pharetra et. Non tellus orci ac auctor augue. Elit at imperdiet dui accumsan sit. Ornare arcu dui vivamus arcu felis. Egestas integer eget aliquet nibh praesent. In hac habitasse platea dictumst quisque sagittis purus. Pulvinar elementum integer enim neque volutpat ac.</p><p>Senectus et netus et malesuada. Nunc pulvinar sapien et ligula ullamcorper malesuada proin. Neque convallis a cras semper auctor. Libero id faucibus nisl tincidunt eget. Leo a diam sollicitudin tempor id. A lacus vestibulum sed arcu non odio euismod lacinia. In tellus integer feugiat scelerisque. Feugiat in fermentum posuere urna nec tincidunt praesent. Porttitor rhoncus dolor purus non enim praesent elementum facilisis. Nisi scelerisque eu ultrices vitae auctor eu augue ut lectus. Ipsum faucibus vitae aliquet nec ullamcorper sit amet risus. Et malesuada fames ac turpis egestas sed. Sit amet nisl suscipit adipiscing bibendum est ultricies. Arcu ac tortor dignissim convallis aenean et tortor at. Pretium viverra suspendisse potenti nullam ac tortor vitae purus. Eros donec ac odio tempor orci dapibus ultrices. Elementum nibh tellus molestie nunc. Et magnis dis parturient montes nascetur. Est placerat in egestas erat imperdiet. Consequat interdum varius sit amet mattis vulputate enim.</p><>Sit amet nulla facilisi morbi tempus. Nulla facilisi cras fermentum odio eu. Etiam erat velit scelerisque in dictum non consectetur a erat. Enim nulla aliquet porttitor lacus luctus accumsan tortor posuere. Ut sem nulla pharetra diam. Fames ac turpis egestas maecenas. Bibendum neque egestas congue quisque egestas diam. Laoreet id donec ultrices tincidunt arcu non sodales neque. Eget felis eget nunc lobortis mattis aliquam faucibus purus. Faucibus interdum posuere lorem ipsum dolor sit.`;
-
-export const DOCUMENT_TERMS_OF_USE_TITLE = "Политика в отношении обработки персональных данных";
-export const DOCUMENT_TERMS_OF_USE_CONTENT = [
-  {
-    title: '1. Общие положения',
-    titleText: 'Настоящая политика обработки персональных данных составлена в соответствии с требованиями Федерального закона от 27.07.2006. №152-ФЗ «О персональных данных» и определяет порядок обработки персональных данных и меры по обеспечению безопасности персональных данных, предпринимаемые ИП Иванов Иван Иванович (далее – Оператор).',
-    texts: [
-      '1.1. Оператор ставит своей важнейшей целью и условием осуществления своей деятельности соблюдение прав и свобод человека и гражданина при обработке его персональных данных, в том числе защиты прав на неприкосновенность частной жизни, личную и семейную тайну.',
-      '1.2. Настоящая политика Оператора в отношении обработки персональных данных (далее – Политика) применяется ко всей информации, которую Оператор может получить о посетителях веб-сайта http://kengurumom.ru.',
-    ]
-  },
-  {
-    title: '2. Основные понятия, используемые в Политике',
-    texts: [
-      '2.1. Автоматизированная обработка персональных данных – обработка персональных данных с помощью средств вычислительной техники;',
-      '2.2. Блокирование персональных данных – временное прекращение обработки персональных данных (за исключением случаев, если обработка необходима для уточнения персональных данных);',
-      '2.3. Веб-сайт – совокупность графических и информационных материалов, а также программ для ЭВМ и баз данных, обеспечивающих их доступность в сети интернет по сетевому адресу http://kengurumom.ru;',
-      '2.4. Информационная система персональных данных — совокупность содержащихся в базах данных персональных данных, и обеспечивающих их обработку информационных технологий и технических средств;',
-      '2.5. Обезличивание персональных данных — действия, в результате которых невозможно определить без использования дополнительной информации принадлежность персональных данных конкретному Пользователю или иному субъекту персональных данных;',
-      '2.6. Обработка персональных данных – любое действие (операция) или совокупность действий (операций), совершаемых с использованием средств автоматизации или без использования таких средств с персональными данными, включая сбор, запись, систематизацию, накопление, хранение, уточнение (обновление, изменение), извлечение, использование, передачу (распространение, предоставление, доступ), обезличивание, блокирование, удаление, уничтожение персональных данных;',
-      '2.7. Оператор – государственный орган, муниципальный орган, юридическое или физическое лицо, самостоятельно или совместно с другими лицами организующие и (или) осуществляющие обработку персональных данных, а также определяющие цели обработки персональных данных, состав персональных данных, подлежащих обработке, действия (операции), совершаемые с персональными данными;',
-      '2.8. Персональные данные – любая информация, относящаяся прямо или косвенно к определенному или определяемому Пользователю веб-сайта  http://kengurumom.ru;',
-      '2.9. Пользователь – любой посетитель веб-сайта http://kengurumom.ru;',
-      '2.10. Предоставление персональных данных – действия, направленные на раскрытие персональных данных определенному лицу или определенному кругу лиц;',
-      '2.11. Распространение персональных данных – любые действия, направленные на раскрытие персональных данных неопределенному кругу лиц (передача персональных данных) или на ознакомление с персональными данными неограниченного круга лиц, в том числе обнародование персональных данных в средствах массовой информации, размещение в информационно-телекоммуникационных сетях или предоставление доступа к персональным данным каким-либо иным способом;',
-      '2.12. Трансграничная передача персональных данных – передача персональных данных на территорию иностранного государства органу власти иностранного государства, иностранному физическому или иностранному юридическому лицу;',
-      '2.13. Уничтожение персональных данных – любые действия, в результате которых персональные данные уничтожаются безвозвратно с невозможностью дальнейшего восстановления содержания персональных данных в информационной системе персональных данных и (или) уничтожаются материальные носители персональных данных.',
-    ]
-  },
-  {
-    title: '3. Оператор может обрабатывать следующие персональные данные Пользователя:',
-    texts: [
-      '3.1. Фамилия, имя, отчество;',
-      '3.2. Электронный адрес;',
-      '3.3. Номера телефонов;',
-      '3.4. Адрес доставки;',
-      '3.5. Также на сайте происходит сбор и обработка обезличенных данных о посетителях (в т.ч. файлов «cookie») с помощью сервисов интернет-статистики (Яндекс Метрика и Гугл Аналитика и других).',
-      '3.6. Вышеперечисленные данные далее по тексту Политики объединены общим понятием Персональные данные.',
-    ]
-  },
-  {
-    title: '4. Цели обработки персональных данных',
-    texts: [
-      '4.1. Цель обработки персональных данных Пользователя — предоставление услуг (оформление заказа, осуществление доставки и т.д.), предоставление доступа Пользователю к сервисам, информации и/или материалам, содержащимся на веб-сайте.',
-      '4.2. Также Оператор имеет право направлять Пользователю уведомления о новых продуктах и услугах, специальных предложениях и различных событиях. Пользователь всегда может отказаться от получения информационных сообщений, направив Оператору письмо на адрес электронной почты kenguru.mom@gmail.com с пометкой «Отказ от уведомлений о новых продуктах и услугах и специальных предложениях».',
-      '4.3. Обезличенные данные Пользователей, собираемые с помощью сервисов интернет-статистики, служат для сбора информации о действиях Пользователей на сайте, улучшения качества сайта и его содержания.',
-    ]
-  },{
-    title: '5. Правовые основания обработки персональных данных',
-    texts: [
-      '5.1. Оператор обрабатывает персональные данные Пользователя только в случае их заполнения и/или отправки Пользователем самостоятельно через специальные формы, расположенные на сайте  http://kengurumom.ru. Заполняя соответствующие формы и/или отправляя свои персональные данные Оператору, Пользователь выражает свое согласие с данной Политикой.',
-      '5.2. Оператор обрабатывает обезличенные данные о Пользователе в случае, если это разрешено в настройках браузера Пользователя (включено сохранение файлов «cookie» и использование технологии JavaScript).',
-    ]
-  },{
-    title: '6. Порядок сбора, хранения, передачи и других видов обработки персональных данных',
-    titleText: 'Безопасность персональных данных, которые обрабатываются Оператором, обеспечивается путем реализации правовых, организационных и технических мер, необходимых для выполнения в полном объеме требований действующего законодательства в области защиты персональных данных.',
-    texts: [
-      '6.1. Оператор обеспечивает сохранность персональных данных и принимает все возможные меры, исключающие доступ к персональным данным неуполномоченных лиц.',
-      '6.2. Персональные данные Пользователя никогда, ни при каких условиях не будут переданы третьим лицам, за исключением случаев, связанных с исполнением действующего законодательства.',
-      '6.3. В случае выявления неточностей в персональных данных, Пользователь может актуализировать их самостоятельно, путем направления Оператору уведомление на адрес электронной почты Оператора kenguru.mom@gmail.com с пометкой «Актуализация персональных данных».',
-      '6.4. Срок обработки персональных данных является неограниченным. Пользователь может в любой момент отозвать свое согласие на обработку персональных данных, направив Оператору уведомление посредством электронной почты на электронный адрес Оператора kenguru.mom@gmail.com с пометкой «Отзыв согласия на обработку персональных данных».',
-    ]
-  },{
-    title: '7. Трансграничная передача персональных данных',
-    titleText: 'Безопасность персональных данных, которые обрабатываются Оператором, обеспечивается путем реализации правовых, организационных и технических мер, необходимых для выполнения в полном объеме требований действующего законодательства в области защиты персональных данных.',
-    texts: [
-      '7.1. Оператор до начала осуществления трансграничной передачи персональных данных обязан убедиться в том, что иностранным государством, на территорию которого предполагается осуществлять передачу персональных данных, обеспечивается надежная защита прав субъектов персональных данных.',
-      '7.2. Трансграничная передача персональных данных на территории иностранных государств, не отвечающих вышеуказанным требованиям, может осуществляться только в случае наличия согласия в письменной форме субъекта персональных данных на трансграничную передачу его персональных данных и/или исполнения договора, стороной которого является субъект персональных данных.',
-    ]
-  },{
-    title: '8. Заключительные положения',
-    texts: [
-      '8.1. Пользователь может получить любые разъяснения по интересующим вопросам, касающимся обработки его персональных данных, обратившись к Оператору с помощью электронной почты kenguru.mom@gmail.com.',
-      '8.2. В данном документе будут отражены любые изменения политики обработки персональных данных Оператором. Политика действует бессрочно до замены ее новой версией.',
-      '8.3. Актуальная версия Политики в свободном доступе расположена в сети Интернет по адресу http://kengurumom.ru/privacy.'
-    ]
-  },
-];
